submissions/bernardo-braga-perobeli/solution/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-jose[cryptography]
python-dotenv
python-multipart
openpyxl
google-genai
transformers
sentence-transformers
torch
pandas
numpy
scikit-learn
joblib
